% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{selecao}
\alias{selecao}
\title{Selecao de códigos CID/SIM}
\usage{
selecao(data, CID)
}
\arguments{
\item{data}{a data.frame.}

\item{CID}{a CID code.}
}
\value{
filter  of \code{data} by \code{cid}.
}
\description{
Selecao de códigos CID/SIM
}
\examples{
meus_dados <-selecao(data=banco,CID='I50')
meus_dados2 <-selecao(banco,CID='I21')
meus_dados3 <-selecao(banco,CID = c("I0[5-8]|I091|I3[4-9]|Q2[2-3]|T820|T822|T826|Z95"))
}
